pagination and review funtionality added

+++ frontend/src/screens/admin/ProductListScreen.jsx
@@ -4,33 +4,32 @@ import { FaEdit, FaPlus, FaTrash } from 'react-icons/fa';
 import { useParams } from 'react-router-dom';
 import Message from '../../components/Message';
 import Loader from '../../components/Loader';
-//import Paginate from '../../components/Paginate';
+import Paginate from '../../components/Paginate';
 import {
   useGetProductsQuery,
- /*  useDeleteProductMutation, */
+  useDeleteProductMutation,
   useCreateProductMutation,
 } from '../../slices/productsApiSlice';
 import { toast } from 'react-toastify';
 
 const ProductListScreen = () => {
   const { pageNumber } = useParams();
-
-  const { data: products, isLoading, error, refetch } = useGetProductsQuery({
+  const { data, isLoading, error, refetch } = useGetProductsQuery({
     pageNumber,
   });
 
-/*   const [deleteProduct, { isLoading: loadingDelete }] =
-    useDeleteProductMutation(); */
+  const [deleteProduct, { isLoading: loadingDelete }] =
+    useDeleteProductMutation();
 
   const deleteHandler = async (id) => {
-  /*   if (window.confirm('Are you sure')) {
+    if (window.confirm('Are you sure')) {
       try {
         await deleteProduct(id);
         refetch();
       } catch (err) {
         toast.error(err?.data?.message || err.error);
       }
-    } */
+    }
   };
 
   const [createProduct, { isLoading: loadingCreate }] =
@@ -79,7 +78,7 @@ const ProductListScreen = () => {
               </tr>
             </thead>
             <tbody>
-              {products.map((product) => (
+              {data.products.map((product) => (
                 <tr key={product._id}>
                   <td>{product._id}</td>
                   <td>{product.name}</td>
@@ -104,6 +103,7 @@ const ProductListScreen = () => {
               ))}
             </tbody>
           </Table>
+          <Paginate pages={data.pages} page={data.page} isAdmin={true}  />
         </>
       )}
     </>
